+++ source/deprecated/isaacsim.sensors.physics/python/impl/effort_sensor.py
@@ -107,7 +107,6 @@ class EffortSensor(SingleArticulation):
             on_event=self._stage_open_callback_fn,
             observer_name="isaacsim.sensors.physics.EffortSensor.initialize._stage_open_callback",
         )
-        timeline = omni.timeline.get_timeline_interface()
         self._timer_reset_callback_stop = carb.eventdispatcher.get_eventdispatcher().observe_event(
             event_name=omni.timeline.GLOBAL_EVENT_STOP,
             on_event=self._timeline_stop_callback_fn,